refactor: Restructure frontend components and remove Summaryview
- Updated `main.jsx` to import `App` from the new components directory.
- Deleted the `Summaryview` component as it is no longer needed.
- Introduced new components including `App`, `Footer`, `Header`, `HeroSection`, `Features`, `Pricing`, `Testimonials`, `Dashboard`, `RecordingList`, `RecordingData`, `UserProfile`, and `UserProfileEdit` to enhance the application structure and user experience.
- Added new styles in `App.css` to support the updated layout and design.
- Implemented authentication and user profile management features in the new components.
